Sketch training implementation
This allows training to run, although it does not yet seem successful.

Unlike the paper, this performs only one training pass on each
subvolume. The training subvolume is the same as the input size, so
there are no moves or feeding back of mask outputs. This enhancement
will be added later.