Volumes: refactor
Refactor volumes to:

- Make partitioning and downsampling operations wrap volumes as partial
  coordinate transform views, so that they are composable and better
  isolated.
- Separate generation of subvolume locations from extraction of
  subvolumes, so that locations can be easily injected from other
  sources, e.g., file input.
- Make use of octree-backed block sparse volumes less intrusive/code
  duplicating than before.
- Downsample volumes based on a target resolution instead of downsample
  zoom levels, which means less global config access and more natural
  handling for already downsampled volumes, e.g., image stack zooms.

The current volume view wrapping implementation is inelegant and
brittle (a unified transform-based view would be preferable), but
functional since this code is in flux. The main constraint on this is
the necessity of masking labels before transformation.

There is a 10% subvolume generation performance regression from these
changes.