Allow output FOV to be smaller than input
This is useful because predictions at the boundary of the output FOV can
often not be undone, making a prediction without sufficient context
often results in error. It's also useful because the network
architecture may not be deep enough to propagate information from one
input FOV edge to the other.

Previously achieved by cropping output in the Region, but this explicit
network implementation can be more efficient by making predictions that
will be discarded.