Regions: align move grid with seed voxel
Rather than align the move grid to the volume coordinate system, align
it so that the seed voxel is a valid grid location. While this is
necessary for cropped output FOV shapes, it will also fix an annoyance
with highly-downsampled volumes where the seed voxel would move outside
the intended process because of grid alignment.