BREAKING: Upgrade to Keras 2 / TF 1.1
This breaks some old saved models. The factors determining which old
models are loadable and which are not are unknown, but models trained
since around the dimension reordering in
b6bec20b155b9446d54172ce29f05a3796c2edab seems to work while those prior
do not, although the breakage is not related to this reordering (or the
tranposition wrapper).

Also requires updates to NumPy and SciPy.

Requires changes to multi GPU wrapper, but at least for small GPU counts
and training sizes performance is unchanged.

Fixes U-net with multi GPU. Seems to fix the bug preventing convergence
with cropped output FOV convolution nets also. Requires U-net depth
reduced to 3 because of a TF dump.

Note that this did not replace the upsampling convolution kludge in
U-net, because Keras does not offer a builtin 3D upsampling/fractional
stride conv layer, so a custom one will need to be written again.