Regions: allow blocks extending outside bounds
It is necessary to allow blocks to extend outside the region bounds so
that filling can fill to the boundaries of the available image data. To
support this, a `block_padding` property is added which can be set to a
valid padding mode for `numpy.pad` to pad input data when it lies
outside the region.

By default `block_padding` is None, which will fail assertion as we
never want to use this padding behavior unexpectedly.