Training: roundrobin over subv gens, not training gens
This drastically reduces memory usage and improve performance when using
many volumes, as there is a single moving training generator and
callback, rather than one per volume.

This also makes it such that subvolumes from multiple volumes will be
mixed in a single batch.

For now, generator resetting is disabled, but will be fixed in a later
commit refactoring roundrobin.