Training: specify number of training gen workers
Training data generation used to be handled by a number of workers equal
to the number of volumes, because this was also the number of subvolume
generators and training data generators. Now that subvolume generators
are grouped into fewer (previously one) training data generators, allow
this number to be specified.