Filling: parallelize seed filling
Parallelize filling so that separate seed regions are processed on each
GPU. This implementation assures that the result is deterministic with
the serial case by only merging segmented bodies into the final
segmentation in the order in which they were originally queued.

There can be wasted work because the seed location of segmented bodies
may have been covered between queueing for work and when the result is
done, in which case it is discarded.

Closes aschampion/diluvian#5.