Training: move epoch reset to callback
This makes epoch reset much more consistent by making the moving
training generator a resettable class and moving the reset action itself
to a callback. Consequently simplifies much of the validation kludge
logic.

Note that while Keras has a queue, it restarts queuing validation data
at the end of each epoch, so this approach should *not* result in excess
queue for the previous epoch being used at the beginning of the
subsequent epoch.