Postprocessing: fix literal edge case
Fix a case where a seed on the edge (bounds) of a volume is remove from
the mask by morphological closing because SciPy does not have border
modes for binary closing. Instead use grayscale closing with a
boundary mode that will not erode mask at the volume boundaries.

This undoes the erroneous fix in
25ac9b06bf3f3ec26f79c56ff31fb8008687d27d, and is a partial reversion of
e07e8e3afaeb36c6b0323b118ec0723e347f4140, prior to which grayscale
closing was used. Ugh.